+++ src/main/java/local/fishingrodcompat/mixin/TackleBoxContainerMixin.java
@@ -9,7 +9,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.Slot;
-import net.neoforged.neoforge.items.SlotItemHandler;
+import net.minecraftforge.items.SlotItemHandler;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -39,10 +39,10 @@ public abstract class TackleBoxContainerMixin extends AbstractContainerMenu {
     }
 
     @Redirect(
-            method = "clicked",
+            method = "m_150399_",
             at = @At(
                     value = "INVOKE",
-                    target = "Lnet/neoforged/neoforge/items/SlotItemHandler;set(Lnet/minecraft/world/item/ItemStack;)V"
+                    target = "Lnet/minecraftforge/items/SlotItemHandler;m_5852_(Lnet/minecraft/world/item/ItemStack;)V"
             ),
             remap = false
     )